requirements.txt
tabula-py
pandas
sklearn
openai
jpype1
scikit-learn
matplotlib